Backend/requirements_sqlite.txt
Django>=4.2,<6.0
djangorestframework
django-cors-headers
Pillow
reportlab
qrcode
gunicorn
whitenoise
djangorestframework-simplejwt
python-decouple
